examples/hello-world-fullstack/backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
PyJWT==2.10.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0